backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements/requirements.in
#
asgiref==3.5.0
    # via django
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
cryptography==37.0.2
    # via -r requirements/requirements.in
dj-database-url==0.5.0
    # via environs
dj-email-url==1.0.5
    # via environs
django==4.0.4
    # via
    #   -r requirements/requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cache-url==3.4.0
    # via environs
django-cors-headers==3.11.0
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements/requirements.in
environs[django]==9.5.0
    # via -r requirements/requirements.in
gunicorn==20.1.0
    # via -r requirements/requirements.in
idna==3.3
    # via requests
marshmallow==3.15.0
    # via environs
packaging==21.3
    # via marshmallow
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyparsing==3.0.8
    # via packaging
python-dotenv==0.20.0
    # via environs
python-json-logger==2.0.2
    # via -r requirements/requirements.in
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via -r requirements/requirements.in
requests==2.27.1
    # via -r requirements/requirements.in
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via -r requirements/requirements.in
urllib3==1.26.9
    # via requests
whitenoise==6.0.0
    # via -r requirements/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
